Solution to Check if a Number is Prime or Not

+++ Check if a Number is Prime or Not@@6788dd63490637cfc44e1ba2/code.c
@@ -6,7 +6,7 @@ void checkPrime(int n){
         printf("Not Prime");
     }
 
-    for(int i=2; i<sqrt(n); i++){
+    for(int i=2; i<(int)sqrt(n); i++){
         if(n%i == 0){
             printf("Not Prime");
         }
